requirements.txt
# requirements.txt
together
python-dotenv
datasets
pandas
numpy
tqdm